+++ Source/ACE.Server/WorldObjects/Container.cs
@@ -805,7 +805,7 @@ namespace ACE.Server.WorldObjects
             if (Biota.PropertiesCreateList == null)
                 return;
 
-            foreach (var item in Biota.PropertiesCreateList.Where(x => x.DestinationType == (sbyte)DestinationType.Contain || x.DestinationType == (sbyte)DestinationType.ContainTreasure))
+            foreach (var item in Biota.PropertiesCreateList.Where(x => x.DestinationType == DestinationType.Contain || x.DestinationType == DestinationType.ContainTreasure))
             {
                 var wo = WorldObjectFactory.CreateNewWorldObject(item.WeenieClassId);
